feat(config): centralize source decoder construction
## What

- Centralize source-decoder construction behind `NewDecoder` and make the source-specific implementations private.
- Move transport projections into their own file and align package documentation with the remaining public decoder API.
- The removal of direct source-decoder constructors and types is intentional.

## Why

- Keep the configuration package focused on its flag-driven decoder entry point and avoid exposing source-selection internals as a supported API.

## Testing

- `make lint package=config` - passed
- `make specs package=config` - passed (52 config tests).
- Broader CI-only validation (security, full specs, fuzzing, benchmarks, and coverage) was not run locally.

AI-assisted-by: [Codex](https://openai.com/codex/)
